feat: 체스판에 기물들을 초기화를 담당하는 클래스 구현
