+++ src/main/java/domain/board/Board.java
@@ -1,53 +1,27 @@
 package domain.board;
 
 import domain.chessgame.Score;
-import domain.piece.Bishop;
 import domain.piece.EmptyPiece;
-import domain.piece.King;
-import domain.piece.Knight;
-import domain.piece.Pawn;
 import domain.piece.Piece;
-import domain.piece.Queen;
-import domain.piece.Rook;
 import domain.position.ColumnDegree;
 import domain.position.Position;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
 
 public class Board {
 
-    private static final int CHESS_BOARD_SIZE = 8;
+    public static final int CHESS_BOARD_SIZE = 8;
     private static final int PAWN_ALLY_COUNT_CONDITION = 2;
     private final Map<Position, Piece> board;
 
     public Board() {
-        board = emptyBoard();
-    }
-
-    private Map<Position, Piece> emptyBoard() {
-        Map<Position, Piece> emptyBoard = new HashMap<>();
-        for (int row = 0; row < CHESS_BOARD_SIZE; row++) {
-            putInitialPiece(emptyBoard, row);
-        }
-        return emptyBoard;
-    }
-
-    private void putInitialPiece(Map<Position, Piece> emptyBoard, int row) {
-        for (int column = 0; column < CHESS_BOARD_SIZE; column++) {
-            emptyBoard.put(new Position(row, column), new EmptyPiece());
-        }
+        board = InitialBoard.emptyBoard();
     }
 
     public void initChessPieces() {
-        board.putAll(King.createInitialKings());
-        board.putAll(Queen.createInitialQueens());
-        board.putAll(Bishop.createInitialBishops());
-        board.putAll(Knight.createInitialKnights());
-        board.putAll(Rook.createInitialRooks());
-        board.putAll(Pawn.createInitialPawns());
+        InitialBoard.initChessPieces(board);
     }
 
     public Map<Position, Piece> getBoard() {
