+++ src/main/java/domain/piece/King.java
@@ -3,7 +3,6 @@ package domain.piece;
 import domain.chessgame.Score;
 import domain.position.Direction;
 import domain.position.Position;
-import java.util.HashMap;
 import java.util.Map;
 
 public class King extends Piece {
@@ -13,13 +12,6 @@ public class King extends Piece {
 
     public King(boolean isBlack) {
         super(NAME, isBlack, SCORE);
-    }
-
-    public static Map<Position, Piece> createInitialKings() {
-        Map<Position, Piece> initialKings = new HashMap<>();
-        initialKings.put(new Position("e8"), new King(true));
-        initialKings.put(new Position("e1"), new King(false));
-        return initialKings;
     }
 
     @Override
