+++ src/main/java/domain/piece/Knight.java
@@ -3,7 +3,6 @@ package domain.piece;
 import domain.chessgame.Score;
 import domain.position.Direction;
 import domain.position.Position;
-import java.util.HashMap;
 import java.util.Map;
 
 public class Knight extends Piece {
@@ -13,15 +12,6 @@ public class Knight extends Piece {
 
     public Knight(boolean isBlack) {
         super(NAME, isBlack, SCORE);
-    }
-
-    public static Map<Position, Piece> createInitialKnights() {
-        Map<Position, Piece> initialKnights = new HashMap<>();
-        initialKnights.put(new Position("b8"), new Knight(true));
-        initialKnights.put(new Position("g8"), new Knight(true));
-        initialKnights.put(new Position("b1"), new Knight(false));
-        initialKnights.put(new Position("g1"), new Knight(false));
-        return initialKnights;
     }
 
     @Override
