+++ src/main/java/domain/piece/Rook.java
@@ -3,7 +3,6 @@ package domain.piece;
 import domain.chessgame.Score;
 import domain.position.Direction;
 import domain.position.Position;
-import java.util.HashMap;
 import java.util.Map;
 
 public class Rook extends Piece {
@@ -13,15 +12,6 @@ public class Rook extends Piece {
 
     public Rook(boolean isBlack) {
         super(NAME, isBlack, SCORE);
-    }
-
-    public static Map<Position, Piece> createInitialRooks() {
-        Map<Position, Piece> initialRooks = new HashMap<>();
-        initialRooks.put(new Position("a8"), new Rook(true));
-        initialRooks.put(new Position("h8"), new Rook(true));
-        initialRooks.put(new Position("a1"), new Rook(false));
-        initialRooks.put(new Position("h1"), new Rook(false));
-        return initialRooks;
     }
 
     @Override
